Add @types/next-server / port @types/next lib (#29724)
* ported next lib to next-server

* lint fixes

* updated exports from next-server

* added maintainers from next to next-server

* removed deprecations from next-server

* added return value `false` to ServerRoute.match

* fixed formatting
