+++ types/next-server/asset.d.ts
@@ -1,0 +1,2 @@
+export function setAssetPrefix(url: string): void;
+export default function(path: string): string;
