+++ types/next-server/config.d.ts
@@ -1,0 +1,6 @@
+export interface RuntimeConfig {
+    serverRuntimeConfig: any;
+    publicRuntimeConfig?: any;
+}
+export function setConfig(configValue: any): void;
+export default function(): RuntimeConfig;
