+++ types/next/head.d.ts
@@ -1,8 +1,2 @@
-import * as React from "react";
-
-export function defaultHead(): JSX.Element[];
-export default class Head extends React.Component {
-    static canUseDOM: boolean;
-    static peek(): Array<React.ReactElement<any>>;
-    static rewind(): Array<React.ReactElement<any>>;
-}
+export * from 'next-server/head';
+export { default } from 'next-server/head';
