+++ types/next/link.d.ts
@@ -1,21 +1,8 @@
-import * as url from "url";
-import * as React from "react";
-import { UrlLike } from "./router";
+import { LinkProps } from 'next-server/link';
 
 // Deprecated
 export type LinkState = LinkProps;
 // End Deprecated
 
-export interface LinkProps {
-    prefetch?: boolean;
-    shallow?: boolean;
-    scroll?: boolean;
-    replace?: boolean;
-    onError?(error: any): void;
-    href?: string | UrlLike;
-    as?: string | UrlLike;
-    passHref?: boolean;
-    children: React.ReactElement<any>;
-}
-
-export default class Link extends React.Component<LinkProps> {}
+export * from 'next-server/link';
+export { default } from 'next-server/link';
